UPDATE data filtering in stitchscan
UPDATE the filtering in the stitchscan function was previously located
before the data was stitched.

That meant that if one dataset had fewer or more values >0, the datasets
could not be stitched.

Further only the first diffractogram was used.

Now the filtering is done post combining of the data.
Also all diffractograms are used (so any tth value where the data is > 0
is now filtered away