Fix two bugs:
Energy is now read from the hdcm or mlm as the smaller absolute value
between the two. For XRF Maps.
This means that XRF mapping can now be done without having a diffraction
file

meshct_MAXIV now works by sizing the scanarea in the correct order
This means that if the scans need to be reduced in size (because one
scan parameter gives steps and another points), it will now first
downsize the fastaxis, and then, if it still doesn't fit, downsize the
slow axis.